requirements.txt
discord.py==2.6.4
pytz==2021.3
